app/requirements.txt
aiofile==3.8.1 ; python_version >= "3.10" and python_version < "4"
aiopath==0.6.11 ; python_version >= "3.10" and python_version < "4.0"
anyio==3.6.2 ; python_version >= "3.10" and python_version < "4.0"
asgiref==3.6.0 ; python_version >= "3.10" and python_version < "4.0"
backoff==2.2.1 ; python_version >= "3.10" and python_version < "4.0"
caio==0.9.11 ; python_version >= "3.10" and python_version < "4"
certifi==2022.12.7 ; python_version >= "3.10" and python_version < "4.0"
charset-normalizer==3.0.1 ; python_version >= "3.10" and python_version < "4"
click==8.1.3 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and platform_system == "Windows"
decorator==5.1.1 ; python_version >= "3.10" and python_version < "4.0"
deprecated==1.2.13 ; python_version >= "3.10" and python_version < "4.0"
fastapi==0.89.1 ; python_version >= "3.10" and python_version < "4.0"
googleapis-common-protos==1.58.0 ; python_version >= "3.10" and python_version < "4.0"
gunicorn==20.1.0 ; python_version >= "3.10" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.10" and python_version < "4.0"
httpcore==0.16.3 ; python_version >= "3.10" and python_version < "4.0"
httpx==0.23.3 ; python_version >= "3.10" and python_version < "4.0"
idna==3.4 ; python_version >= "3.10" and python_version < "4"
jsonpath-ng==1.5.3 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-api==1.15.0 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-exporter-otlp-proto-http==1.15.0 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-instrumentation-asgi==0.36b0 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-instrumentation-fastapi==0.36b0 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-instrumentation-httpx==0.36b0 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-instrumentation==0.36b0 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-proto==1.15.0 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-sdk==1.15.0 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-semantic-conventions==0.36b0 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-util-http==0.36b0 ; python_version >= "3.10" and python_version < "4.0"
ply==3.11 ; python_version >= "3.10" and python_version < "4.0"
protobuf==4.21.12 ; python_version >= "3.10" and python_version < "4.0"
pydantic==1.10.4 ; python_version >= "3.10" and python_version < "4.0"
python-json-logger==2.0.4 ; python_version >= "3.10" and python_version < "4.0"
requests==2.28.2 ; python_version >= "3.10" and python_version < "4"
rfc3986[idna2008]==1.5.0 ; python_version >= "3.10" and python_version < "4.0"
setuptools==67.1.0 ; python_version >= "3.10" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.10" and python_version < "4.0"
starlette==0.22.0 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.4.0 ; python_version >= "3.10" and python_version < "4.0"
uplink==0.9.7 ; python_version >= "3.10" and python_version < "4.0"
uritemplate==4.1.1 ; python_version >= "3.10" and python_version < "4.0"
urllib3==1.26.14 ; python_version >= "3.10" and python_version < "4"
uvicorn==0.20.0 ; python_version >= "3.10" and python_version < "4.0"
wrapt==1.14.1 ; python_version >= "3.10" and python_version < "4.0"
